Refactored drawerNavItem
