+++ lib/MainDrawer.dart
@@ -10,6 +10,7 @@ import 'package:provider/provider.dart';
 import 'About.dart';
 
 import 'Providers/ThemeProvider.dart';
+import 'UI Components/Drawer_Nav_Item.dart';
 
 class MainDrawer extends StatelessWidget {
   @override
@@ -172,31 +173,6 @@ class MainDrawer extends StatelessWidget {
         ),
       ),
     );
-  }
-}
-
-class drawerNavItem extends StatelessWidget {
-
-  final VoidCallback callback;
-  final IconData  iconData;
-  final String navItemTitle;
-
-  drawerNavItem({
-    this.callback,
-    this.iconData,
-    this.navItemTitle,
-    Key key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return ListTile(
-        onTap: callback,
-        leading:  Icon(iconData),
-        title: Text(
-          navItemTitle,
-          style: TextStyle(fontSize: 18),
-        ));
   }
 }
